Moved ManagedConfigManager to a separate package
