+++ collect_app/src/main/java/org/odk/collect/android/mainmenu/MainMenuActivity.kt
@@ -8,8 +8,8 @@ import org.odk.collect.android.R
 import org.odk.collect.android.activities.ActivityUtils
 import org.odk.collect.android.activities.CrashHandlerActivity
 import org.odk.collect.android.activities.FirstLaunchActivity
-import org.odk.collect.android.application.initialization.ManagedConfigManager
 import org.odk.collect.android.injection.DaggerUtils
+import org.odk.collect.android.mdm.ManagedConfigManager
 import org.odk.collect.android.projects.ProjectSettingsDialog
 import org.odk.collect.android.utilities.ThemeUtils
 import org.odk.collect.androidshared.ui.FragmentFactoryBuilder
